refactor(app): split code into separate files
spplited code into different files so it can be easier to read, added a few more features as well
