--- src/App.css
@@ -1,6 +1,0 @@
-.mapDiv {
-  padding: 0;
-  margin: 0;
-  height: 100%;
-  width: 100%;
-}
