+++ src/envVariables/index.js
@@ -1,0 +1,1 @@
+export const apieKey = process.env.REACT_APP_API_KEY;
