+++ src/index.css
@@ -15,4 +15,11 @@ body {
   -moz-osx-font-smoothing: grayscale;
 }
 
+.mapDiv {
+  padding: 0;
+  margin: 0;
+  height: 100%;
+  width: 100%;
+}
+
 @import "@arcgis/core/assets/esri/themes/dark/main.css";
